moving app routes to sepatate file
